Moved theme into separate component
